refactor: логика маппинга и валидации входных данных от запросов клиента

+++ src/TaskFlux.Network/Packets/AuthorizationRequestPacket.cs
@@ -24,7 +24,7 @@ public class AuthorizationRequestPacket : Packet
         {
             var memory = buffer.AsMemory(0, size);
             var writer = new MemoryBinaryWriter(memory);
-            writer.Write(PacketType.AuthorizationRequest);
+            writer.Write(( byte ) PacketType.AuthorizationRequest);
             AuthorizationMethod.Serialize(ref writer);
             await stream.WriteAsync(memory, token);
         }
